Extend radius of settlements.
Most Indian settlements have a radius of one tile, capitals have a
radius of two tiles. Aztec and Inca settlements have a radius of two
tiles, and their capitals have a radius of three tiles. The radius of
settlements and capitals are now fields of the NationType class, and
attributes of the IndianNationType in the specification.


git-svn-id: https://svn.freecol.org/svnroot/freecol/freecol/trunk@6363 04c0ca24-bb37-0410-963f-fb4f4e597ea7