Prevent ships being classed as persons.
This is a bit curly because isPerson is allowing the foundColony
ability to signify a person for backward compatibility reasons.
We will lose this in due course (and just rely on model.ability.person)
but ATM ships are picking up the Player-level foundColony (which is true
for Europeans), so they need to be explicitly told they are not able
to found colonies --- Unit.hasAbility sorts this out, but it is a bit
non-obvious, hence this long comment.

git-svn-id: https://svn.freecol.org/svnroot/freecol/freecol/trunk@9275 04c0ca24-bb37-0410-963f-fb4f4e597ea7