Put distinct minimumColonySize modifier onto each stockade-type, BR#3522055.
The colonyPanel.minimumColonySize message uses the modifier source to
describe the building that is preventing the size reduction.  The
modifier was only on model.building.stockade, which while correctly
inherited by model.building.fort, then when trying to reduce the
population of a fort, the message complains about a stockade.

git-svn-id: https://svn.freecol.org/svnroot/freecol/freecol/trunk@9783 04c0ca24-bb37-0410-963f-fb4f4e597ea7